requirements.txt
pandas
pyarrow
requests
python-dotenv
boto3